break #7 slice 7b-ii: render each numeric wrapper as its real Go type — closes the narrow const-global residual
pp_type now lowers every numint wrapper to its faithful Go type (GoU8→uint8 … GoI32→int32,
GoI64→int64, GoU64→uint64) instead of all→int64. This makes a narrow-typed binding box
faithfully on its own: `var Uc_100_u8 uint8 = (100 & 0xff)` (was int64), so `any uc_100_u8`
now carries Go dynamic type uint8 — the residual 7b-core deferred.

No widening cascade needed: pp_type only affects explicit type ANNOTATIONS, and the only
emitted narrow annotation is uc_100_u8's var decl (its initializer 100 fits uint8). The
narrow OPS still emit int64+mask EXPRESSIONS (untyped — no annotation forces a type), each
converted to its narrow type at the `any` box by 7b-core. GoI64/GoU64 annotations are
unchanged (int64/uint64).

Golden BYTE-IDENTICAL (full build compiles + runs). The narrow-type layer of break #7 is
now fully faithful: every boxed GoU8…GoI32 value — op-result OR binding — has its real Go
interface identity. Remaining #7: {TInt64,TI64} both box int64 (slice 7c, retire TInt64),
then the tag_runtime_agrees forcing theorem (7d).

+++ main.go
@@ -309,7 +309,7 @@ var Uc_product int64 = 1000000000000
 
 var Uc_100_i64 int64 = 100
 
-var Uc_100_u8 int64 = (100 & 0xff)
+var Uc_100_u8 uint8 = (100 & 0xff)
 
 var Uc_u64_hi uint64 = 9223372036854775808
